+++ components/form-builder/layout/Template.tsx
@@ -1,15 +1,7 @@
 import React, { ReactElement } from "react";
 import Head from "next/head";
 import { useTranslation } from "next-i18next";
-import { serverSideTranslations } from "next-i18next/serverSideTranslations";
-import { GetServerSideProps } from "next";
-import { unstable_getServerSession } from "next-auth";
-import { NextPageWithLayout } from "../_app";
-import { getFullTemplateByID } from "@lib/templates";
-import { FormRecord } from "@lib/types";
 import { useRouter } from "next/router";
-import { authOptions } from "@pages/api/auth/[...nextauth]";
-import { AccessControlError, createAbility } from "@lib/privileges";
 import SkipLink from "@components/globals/SkipLink";
 import Footer from "@components/globals/Footer";
 import Loader from "@components/globals/Loader";
@@ -19,12 +11,7 @@ import {
   NavigationStoreProvider,
   TemplateStoreProvider,
 } from "@components/form-builder/store";
-import { LeftNavigation, Header, Layout } from "@components/form-builder/layout/";
-
-export type PageProps = {
-  tab: string;
-  initialForm: FormRecord | null;
-};
+import { LeftNavigation, Header } from "@components/form-builder/layout/";
 
 export const Template = ({ page }: { page: ReactElement }) => {
   return (
@@ -97,71 +84,3 @@ export const PageTemplate = ({
     <Loader message={t("loading")} />
   );
 };
-
-const Page: NextPageWithLayout<PageProps> = () => {
-  return <Layout />;
-};
-
-Page.getLayout = (page: ReactElement) => {
-  return <Template page={page} />;
-};
-
-export const getServerSideProps: GetServerSideProps = async ({
-  query: { params },
-  locale,
-  req,
-  res,
-}) => {
-  const [tab = "start", formID = null] = params || [];
-
-  const FormbuilderParams: { tab: string; initialForm: null | FormRecord } = {
-    tab,
-    initialForm: null,
-  };
-
-  const session = await unstable_getServerSession(req, res, authOptions);
-
-  if (session && !session.user.acceptableUse) {
-    // If they haven't agreed to Acceptable Use redirect to policy page for acceptance
-    return {
-      redirect: {
-        destination: `/${locale}/auth/policy`,
-        permanent: false,
-      },
-    };
-  }
-
-  // @todo - look into better way to handle setting page for "real" pages
-  if (req.url === "/form-builder/settings") {
-    FormbuilderParams.tab = "settings";
-  }
-
-  if (req.url === "/form-builder/edit") {
-    FormbuilderParams.tab = "create";
-  }
-
-  if (formID && session) {
-    try {
-      const ability = createAbility(session.user.privileges);
-      FormbuilderParams.initialForm = await getFullTemplateByID(ability, formID);
-    } catch (e) {
-      if (e instanceof AccessControlError) {
-        return {
-          redirect: {
-            destination: `/${locale}/admin/unauthorized`,
-            permanent: false,
-          },
-        };
-      }
-    }
-  }
-
-  return {
-    props: {
-      ...FormbuilderParams,
-      ...(locale && (await serverSideTranslations(locale, ["common", "form-builder"]))),
-    },
-  };
-};
-
-export default Page;
