+++ components/form-builder/layout/index.ts
@@ -2,6 +2,7 @@ export { Settings } from "./Settings";
 export { LeftNavigation } from "./LeftNavigation";
 export { PreviewNavigation } from "./PreviewNavigation";
 export { EditNavigation } from "./EditNavigation";
+export { Template, PageTemplate } from "./Template";
 export { Layout } from "./Layout";
 export { Header } from "./Header";
 export { Start } from "./Start";
