+++ pages/form-builder/preview.tsx
@@ -1,9 +1,14 @@
 import React, { ReactElement } from "react";
 import { useTranslation } from "next-i18next";
-
 import { NextPageWithLayout } from "../_app";
-import { Template, PageProps, PageTemplate, getServerSideProps } from "./[[...params]]";
-import { Preview, PreviewNavigation } from "@components/form-builder/layout/";
+import { PageProps } from "@lib/types";
+import { getServerSideProps } from "./index";
+import {
+  Preview,
+  PreviewNavigation,
+  Template,
+  PageTemplate,
+} from "@components/form-builder/layout/";
 
 const Page: NextPageWithLayout<PageProps> = () => {
   const { t } = useTranslation("form-builder");
