+++ pages/form-builder/test-data-delivery.tsx
@@ -2,8 +2,14 @@ import React, { ReactElement } from "react";
 import { useTranslation } from "next-i18next";
 
 import { NextPageWithLayout } from "../_app";
-import { Template, PageProps, PageTemplate, getServerSideProps } from "./[[...params]]";
-import { PreviewNavigation, TestDataDelivery } from "@components/form-builder/layout/";
+import { PageProps } from "@lib/types";
+import { getServerSideProps } from "./index";
+import {
+  PreviewNavigation,
+  TestDataDelivery,
+  PageTemplate,
+  Template,
+} from "@components/form-builder/layout/";
 
 const Page: NextPageWithLayout<PageProps> = () => {
   const { t } = useTranslation("form-builder");
